rag-chatbot/requirements.txt
langchain
openai
pinecone-client
langchain-pinecone
langchain-openai
langchain-community
python-dotenv
pypdf
faiss-cpu
"unstructured[all]"